// src/types/types.ts
export type StatesType = {
    '0'?: number
}
export type ParsType = {
    val?: number
    fun?: null
}
export type BloksType = {
    name: string
    inpN: number
    outpN: number
    pars: ParsType
    states: StatesType
}
export type FilesType = {
    user: string
    name: string
    ser: string
}